added hooks specify

+++ pkg/models/Breed.go
@@ -3,6 +3,6 @@ package models
 type Breed struct {
 	Id          uint   `json:"id"`
 	Name        string `json:"name"`
-	GlobalPrice uint   `json:"global_price"`
+	GlobalPrice int16  `json:"global_price"`
 	Wool        string `json:"wool"`
 }
